requirements.txt
openai==1.30.1
httpx==0.27.2
tiktoken==0.7.0
python-dotenv==1.0.1
pytest==8.2.0
ruff==0.4.4
